% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{clean_string}
\alias{clean_string}
\title{Cleans a string that is expected to be an alphanumeric name.}
\usage{
clean_string(str)
}
\arguments{
\item{str}{String to be cleaned.}
}
\value{
A character vector with the cleaned string.
}
\description{
This function takes a string as input and cleans it by removing any
non-alphanumeric characters and whitespace followed by a forward slash.
The resulting string is then trimmed of leading and trailing whitespace.
}
\examples{
clean_string("Austria [ref]")
#> [1] "Austria"

clean_string("RTBF / VRT")
#> [1] "RTBF, VRT"

clean_string("ARD (NDR)")
#> [1] "ARD (NDR)"

}
