% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescraper-package.R
\docType{package}
\name{rescraper-package}
\alias{rescraper}
\alias{rescraper-package}
\title{rescraper: Scrapes Eurovision Data}
\description{
Scrapes Eurovision Song Contest (ESC) data across several categories and sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phiido/rescraper}
  \item Report bugs at \url{https://github.com/phiido/rescraper/issues}
}

}
\author{
\strong{Maintainer}: Thomas Roosdorp \email{thomas@roosdorp.se} (\href{https://orcid.org/0009-0002-5146-2777}{ORCID}) [copyright holder]

}
\keyword{internal}
