% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_contests}
\alias{scrape_contests}
\title{Scrape information about the Eurovision Song Contest events}
\usage{
scrape_contests(years = resc::get_years()$all)
}
\arguments{
\item{years}{A vector of years for which to scrape information. Defaults to
all available years using \code{\link[resc:get_years]{resc::get_years()}}.}
}
\value{
A tibble containing the scraped information. The tibble has the
following columns:
\itemize{
\item \code{year}: The year of the contest.
\item \code{edition}: The contest edition.
\item \code{slogan}: The contest slogan
\item \code{url_logo}: URL of the contest logo.
\item \code{country_code}: Country code for the hosting country
\item \code{city}: The city where the contest was held.
\item \code{venue}: The venue that the contest was held at
\item \code{presenters}: Names of presenters
\item \code{musical_director}: Name of musical director
\item \code{director}: Name of director
\item \code{executive_producers}: Names of executive producers
\item \code{executive_supervisors}: Names of executive supervisors
}
}
\description{
This function fetches information about the Eurovision Song Contest events
for the specified years. If no years are specified, all available years
are used.
}
