% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_countries}
\alias{scrape_countries}
\title{Scrapes info about countries in Eurovision Song Contest}
\usage{
scrape_countries()
}
\value{
A tibble containing the scraped information.
}
\description{
This function fetches information about countries that have participated or still participate in
the Eurovision Song Contest as an entry.
}
\examples{
scrape_countries()

}
