% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_voting_systems}
\alias{scrape_voting_systems}
\title{Scrapes the voting system information from the Eurovision Song Contest.}
\usage{
scrape_voting_systems()
}
\value{
A named list with the voting system rules and tie breaking rules.
}
\description{
This function scrapes the voting system information from the Eurovision
Song Contest wiki page. The information is returned as a \code{list} with the
following columns:
}
\details{
\itemize{
\item \code{year_start}: Starting year for ruleset.
\item \code{year_end}: End year for ruleset.
\item \code{point_allocation}: How points are allocated from each country.
\item \code{use}: How the tie breaking rule is used.
\item \code{description}: A description of the ruleset.
}
}
