% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_countries}
\alias{scrape_countries}
\title{Scrape countries have participated or still participates in Eurovision as an entry}
\usage{
scrape_countries(quiet = FALSE)
}
\arguments{
\item{quiet}{Should scraping be quiet and not output to console}
}
\value{
A tibble
}
\description{
Scrape countries have participated or still participates in Eurovision as an entry
}
